+++ changelog.md
@@ -7,6 +7,8 @@ Bump in TOC file for Season of Discovery and Hardcore update (Classic Era)
 Shout-out to abecks for his contribution to the code. Thanks!
 
 New runes for Season of Discovery:
+- New SAO: Mage's Hot Streak
+- New GAB: Mage's Pyroblast button glows during Hot Streak
 - New SAO: Mage's Missile Barrage, blue-ish tint to differ from Arcane Blast
 - New GAB: Mage's Arcane Missiles button glows during Missile Barrage